moved two more test-classe from search-old into search-api
