+++ applications/search-api/src/test/java/no/dcat/portal/query/ModelFormatterTest.java
@@ -1,4 +1,4 @@
-package no.dcat.portal.webapp;
+package no.dcat.portal.query;
 
 import org.apache.jena.rdf.model.Model;
 import org.apache.jena.rdf.model.ModelFactory;
