+++ applications/search-api/src/test/java/no/dcat/portal/query/SupportedFormatTest.java
@@ -1,4 +1,4 @@
-package no.dcat.portal.webapp;
+package no.dcat.portal.query;
 
 import org.apache.jena.riot.Lang;
 import org.junit.Test;
